forwardmove and sidemove of ticcmd_t needs to be explicitly signed in order for them to operate correctly.
git-svn-id: svn://svn.code.sf.net/p/hhexen/code/trunk/hhexen@538 989aa981-fe4f-0410-aa6a-c19d413bcdcd

+++ base/g_game.c
@@ -1700,8 +1700,8 @@ static void G_WriteDemoTiccmd (ticcmd_t *cmd)
 {
 	if (gamekeydown['q'])		// press q to end demo recording
 		G_CheckDemoStatus ();
-	*demo_p++ = cmd->forwardmove;
-	*demo_p++ = cmd->sidemove;
+	*demo_p++ = (byte) cmd->forwardmove;
+	*demo_p++ = (byte) cmd->sidemove;
 	*demo_p++ = cmd->angleturn>>8;
 	*demo_p++ = cmd->buttons;
 	*demo_p++ = cmd->lookfly;
